refactor(store): split stream.rs into stream/{cursor,progress,decoder}
Single-file stream.rs had grown to 1011 LOC across four distinct
concerns. Split into a module dir with three contextual files:

- stream/cursor.rs — EventStream trait, EventStreamExt combinators
  (fold/scan/decoder entry), Disposition, and the internal shutdown-
  bias select helper.
- stream/progress.rs — Progress accumulator + Step outcome enum used
  by try_scan_until, plus their inline sync tests.
- stream/decoder.rs — DecoderBuilder typestate (5 markers), the two
  terminal wrappers DecodedStream and BorrowedDecodedStream.
- stream/mod.rs — module declarations + the same public re-export
  surface (Disposition, EventStream, EventStreamExt, Progress, Step,
  DecoderBuilder, NeedsCodec, NoUpcaster, WithCodec, WithBorrowing-
  Codec, WithUpcaster, DecodedStream, BorrowedDecodedStream).

History preserved via git rename for the largest piece
(stream.rs → stream/cursor.rs). Selected and select_next_or_shutdown
promoted from private to pub(super) so decoder.rs can reuse the
same shutdown-bias machinery.

No public API change. All external callers (`nexus_store::stream::*`
paths in tests, benches, and downstream crates) resolve through the
preserved re-exports.
